+++ src/components/MultiSelect.tsx
@@ -1,17 +1,20 @@
 import React, { useRef, useState } from 'react';
 import { Text, TouchableOpacity, View, useColorScheme } from 'react-native';
 import styles from '../styles';
-import type { Data, SelectorPos, MultiSelectProperties } from '../types';
+import type { Data, SelectorRect, MultiSelectProperties } from '../types';
 import SelectionList from './SelectionList';
 
 /* Renders a multi-selector component. Takes in props defined in the MultiSelectProperties type. */
 const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
   const style = styles[useColorScheme() === 'dark' ? 1 : 0];
   const ref = useRef<TouchableOpacity>(null);
-  const [listWidth, setListWidth] = useState<string | number>(props.listStyle?.width ?? 0);
-  const [listX, setListX] = useState<number>(0);
   const [listDisplay, setListDisplay] = useState<boolean>(false);
-  const [pos, setPos] = useState<SelectorPos>({top: 0, bottom: 0});
+  const [refRect, setRefRect] = useState<SelectorRect>({
+    x: 0,
+    y: 0,
+    width: 0,
+    height: 0,
+  });
   const [selected, setSelected] = useState<Data[]>([]);
   const selectItem = (items: Data[]) => {
     setSelected(items);
@@ -21,14 +24,13 @@ const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
     ...data.filter((d: Data) => d.priority),
     ...data.filter((d: Data) => !d.priority),
   ];
-  const updatePos = (display = false) => 
+  const updatePos = (display = false) =>
     ref.current?.measureInWindow((x, y, width, height) => {
-      setListX(x);
-      if (props.listStyle?.width === undefined)
-        setListWidth(width);
-      setPos({
-        top: y - (props.listHeight ?? 200) - 5,
-        bottom: y + height + 5,
+      setRefRect({
+        x: x,
+        y: y - 5,
+        width: props.boxStyle?.width ?? width,
+        height: height + 10,
       });
       if (display)
         setListDisplay(true);
@@ -91,12 +93,10 @@ const MultiSelect = (props: MultiSelectProperties): React.JSX.Element => {
         onSelect={selectItem}
         selected={selected}
         clearSelected={() => setSelected([])}
-        listX={listX}
-        listWidth={listWidth}
         listHeight={props.listHeight ?? 200}
         display={listDisplay}
         hide={() => setListDisplay(false)}
-        selectorPos={pos}
+        selectorRect={refRect}
       />
     </View>
   );