+++ src/types.d.ts
@@ -10,9 +10,11 @@ export interface Data {
   data?: object;
 }
 
-export interface SelectorPos {
-  top: number;
-  bottom: number;
+export interface SelectorRect {
+  x: number;
+  y: number;
+  width: string | number;
+  height: number;
 }
 
 export interface ListProperties {
@@ -28,12 +30,10 @@ export interface ListProperties {
   onSelect: ((e: Data) => void) | ((e: Data[]) => void);
   selected: Data[] | Data;
   clearSelected?: () => void;
-  listX: number;
-  listWidth: string | number;
   listHeight: number;
   display: boolean;
   hide: () => void;
-  selectorPos: SelectorPos;
+  selectorRect: SelectorRect;
 }
 
 export interface MultiSelectProperties {