+++ CHANGELOG.md
@@ -16,13 +16,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Internal: Replaced `SelectorPos`, `listWidth`, and `listX` with `SelectorRect` ([#21](https://github.com/rhventures/react-native-dropdown-selector/pull/21))
+
 ### Removed
 
 ### Fixed
 
 - Width of the dropdown lists becomes equal with the selector box by default, rather than expanding as wide as the screen width regardless of the selector box width ([#20](https://github.com/rhventures/react-native-dropdown-selector/pull/20)).
-- Setting a custom `width` style for `listStyle` now works as intended ([#20](https://github.com/rhventures/react-native-dropdown-selector/pull/20)).
+- Setting a custom numeric `width` style for `listStyle` now works as intended ([#20](https://github.com/rhventures/react-native-dropdown-selector/pull/20)).
 - Setting `center` for a custom `alignSelf` style for `listStyle` will now center it correctly ([#20](https://github.com/rhventures/react-native-dropdown-selector/pull/20)).
+- Specifying a percentage for a custom `width` style for `listStyle` will now be properly handled ([#21](https://github.com/rhventures/react-native-dropdown-selector/pull/21))
 
 ## [0.1.0] - 2024-06-28